% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPPV_frame.R
\name{VPPV_frame}
\alias{VPPV_frame}
\title{Making a VPPV data frame for further analysis}
\usage{
VPPV_frame(data, time, RR0, RR1, PP, MVR)
}
\arguments{
\item{data}{data.frame or tibble containing the raw data}

\item{time}{column name of the variable that contains the timestamps of the ECG-Rwave of each individual beat (in seconds)}

\item{RR0}{column name of the variable that contains the length of the RR0 intervals (in seconds)}

\item{RR1}{column name of the variable that contains the length of the RR1 intervals (in seconds)}

\item{PP}{column name of the variable that contains the value of the Pulse Pressure of each individual beat in mmHg}

\item{MVR}{value of the mechanical ventilation rate expressed in cycles per minute}
}
\value{
dataset with correct column names and calculated relative timing of each individual beat for further analysis
with the functions of the vppv-package.
}
\description{
Making a VPPV data frame for further analysis
}
\examples{
b <- afib
colnames(b) <- c('TIME', 'rr0','rr1','pp')
## assuming a MVR of 12 cycles per minute
VPPV_frame(b, TIME,rr0,rr1,pp, 12)
}
