% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPPV_rep.R
\name{VPPV_rep}
\alias{VPPV_rep}
\title{Replicates of VPPV_value using a Metropolis Hastings sample}
\usage{
VPPV_rep(model_gam)
}
\arguments{
\item{model_gam}{A mgcv::gam model resulting from the VPPV_gam function}
}
\value{
a data frame with 1000 values of VPPV based on resampling the coeficients of the model a 1000 times
}
\description{
Replicates of VPPV_value using a Metropolis Hastings sample
}
\examples{
dat = VPPV_frame(afib, time, RR0,RR1, PP,12)
model_gam <- VPPV_gam(dat)
VPPV_value(model_gam)
dist_gam <- VPPV_rep(model_gam)
quantile(dist_gam, c(0.027,0.5,0.975),na.rm=TRUE)
plot(density(dist_gam$VPPV))
}
