% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPPV_gam.R
\name{VPPV_gam}
\alias{VPPV_gam}
\title{Making a GAM model}
\usage{
VPPV_gam(Dat)
}
\arguments{
\item{Dat}{Dat is a VPPV data frame that contains the following columns : time,RR0, RR1, timing and PP}
}
\value{
mgcv model
}
\description{
Making a GAM model
}
